+++ src/aux/rectangle.rs
@@ -2,17 +2,17 @@ use std::clone::*;
 
 #[derive(Copy, Clone)]
 pub struct Rectangle {
-    x: f64,
-    y: f64,
-    width: f64,
-    height: f64,
+    x: i32,
+    y: i32,
+    width: i32,
+    height: i32,
 }
 
 impl Rectangle {
-    pub fn new(x: f64, y: f64, width: f64, height: f64) -> Self {
+    pub fn new(x: i32, y: i32, width: i32, height: i32) -> Self {
         Self { x, y, width, height }
     }
-    pub fn mirror(&self, width: f64) -> Self {
+    pub fn mirror(&self, width: i32) -> Self {
         Rectangle::new(width - self.x - self.width, self.y, self.width, self.height)
     }
 
@@ -27,26 +27,26 @@ impl Rectangle {
     }
 
     pub fn intersects(&self, other: &Rectangle) -> bool {
-        self.intersect(other).width > 0.0 && self.intersect(other).height > 0.0
+        self.intersect(other).width > 0 && self.intersect(other).height > 0
     }
 
-    pub fn includes(&self, x: f64, y: f64) -> bool {
+    pub fn includes(&self, x: i32, y: i32) -> bool {
         self.x <= x && x < self.x + self.width && self.y <= y && y < self.y + self.height
     }
 
-    pub fn chop_start_h(&self, x: f64) -> Rectangle {
+    pub fn chop_start_h(&self, x: i32) -> Rectangle {
         Rectangle::new(self.x, self.y, x - self.x, self.height)
     }
 
-    pub fn chop_end_h(&self, x: f64) -> Rectangle {
+    pub fn chop_end_h(&self, x: i32) -> Rectangle {
         Rectangle::new(x, self.y, self.width - x + self.x, self.height)
     }
 
-    pub fn chop_start_v(&self, y: f64) -> Rectangle {
+    pub fn chop_start_v(&self, y: i32) -> Rectangle {
         Rectangle::new(self.x, self.y, self.width, y - self.y)
     }
 
-    pub fn chop_end_v(&self, y: f64) -> Rectangle {
+    pub fn chop_end_v(&self, y: i32) -> Rectangle {
         Rectangle::new(self.x, y, self.width, self.height - y + self.y)
     }
 
@@ -55,19 +55,19 @@ impl Rectangle {
 
         for c in pattern.chars() {
             if c == 't' {
-                result = Rectangle::new(self.x, self.y, self.width, self.height/2.0)
+                result = Rectangle::new(self.x, self.y, self.width, self.height/2)
             }
 
             if c == 'b' {
-                result = Rectangle::new(self.x, self.y + self.height/2.0, self.width, self.height/2.0)
+                result = Rectangle::new(self.x, self.y + self.height/2, self.width, self.height/2)
             }
 
             if c == 's' {
-                result = Rectangle::new(self.x, self.y, self.width/2.0, self.height)
+                result = Rectangle::new(self.x, self.y, self.width/2, self.height)
             }
 
             if c == 'e' {
-                result = Rectangle::new(self.x + self.width/2.0, self.y, self.width/2.0, self.height)
+                result = Rectangle::new(self.x + self.width/2, self.y, self.width/2, self.height)
             }
         }
 
@@ -77,24 +77,24 @@ impl Rectangle {
     pub fn center(&self, other: &Rectangle) -> Rectangle {
         let horizontal_surplus = self.width - other.width;
         let vertical_surplus = self.height - other.height;
-        let x = self.x + horizontal_surplus /2.0 - other.x;
-        let y = self.y + vertical_surplus /2.0 - other.y;
+        let x = self.x + horizontal_surplus /2 - other.x;
+        let y = self.y + vertical_surplus /2 - other.y;
         Rectangle::new(x, y, other.width, other.height)
     }
 
-    pub fn x(&self) -> f64 {
+    pub fn x(&self) -> i32 {
         self.x
     }
 
-    pub fn y(&self) -> f64 {
+    pub fn y(&self) -> i32 {
         self.y
     }
 
-    pub fn width(&self) -> f64 {
+    pub fn width(&self) -> i32 {
         self.width
     }
 
-    pub fn height(&self) -> f64 {
+    pub fn height(&self) -> i32 {
         self.height
     }
 }
